Add Privacy Policy & TOS
The use of {% verbatim %} here makes the Twig code appear literally and does not actually parse the Twig tags. This prevents the display of links to the website Privacy Policy & TOS. I have assumed you have already added `footer` as a global variable, if not please update the commit accordingly. 

Thanks!